+++ src/main/java/com/ycourlee/ms/labbooking/enums/EResourceType.java
@@ -12,11 +12,11 @@ public enum EResourceType {
     /**
      * 1.菜单 2.API
      */
-    MENU(((byte) 1), "菜单"),
-    API(((byte) 2), "API"),
+    MENU(1, "菜单"),
+    API(2, "API"),
     ;
 
-    private static final Map<Byte, EResourceType> ALL = new HashMap<Byte, EResourceType>();
+    private static final Map<Integer, EResourceType> ALL = new HashMap<Integer, EResourceType>();
 
     static {
         for (EResourceType item : EResourceType.values()) {
@@ -24,23 +24,23 @@ public enum EResourceType {
         }
     }
 
-    private final byte   code;
-    private final String name;
+    private final Integer code;
+    private final String  name;
 
-    EResourceType(byte code, String name) {
+    EResourceType(Integer code, String name) {
         this.code = code;
         this.name = name;
     }
 
-    public static EResourceType instanceOf(int code) {
+    public static EResourceType instanceOf(Integer code) {
         return ALL.get(code);
     }
 
-    public static String getNameByCode(int code) {
+    public static String getNameByCode(Integer code) {
         return Objects.nonNull(instanceOf(code)) ? instanceOf(code).getName() : "";
     }
 
-    public byte getCode() {
+    public Integer getCode() {
         return code;
     }
